[refactor] LostAnimal에 대한 UseCase를 만들어서 분리
