+++ src/main/java/kr/co/pawong/pwbe/lostPost/application/port/in/QueryLostPostDataUseCase.java
@@ -11,6 +11,4 @@ public interface QueryLostPostDataUseCase {
 
     LostPostDetailResponse findLostPostById(Long lostPostId);
 
-    List<LostPostCard> getLostAnimalsByIds(List<LostAnimalQuery> lostAnimalQueries);
-
 }
